requirements.txt
Flask>=2.2.0
requests>=2.31.0
google-generativeai>=0.3.0
gunicorn

